refact: move controllers to controllers-factory
