+++ src/input/controllers-factory/users/logout-user-controller.ts
